+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+## [0.11.0] - 2026-08-31
+
 ### Added
 
 - Binary-safe hashes through `HSET`, `HSETNX`, `HGET`, `HMGET`, `HGETALL`,
@@ -261,7 +263,8 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 - Keys cannot contain whitespace, and values are Unicode strings rather than
   binary-safe byte sequences.
 
-[Unreleased]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ivan-gromov-dev/rustydb/compare/v0.7.0...v0.8.0